First checkpoint
I have get from server and local files setup. I have post to server for headset registration setup, but I've got a lot of cleanup to do. This is just after I got the REST requests working, but there's a lot I still don't understand about the POST request and how UnityWebRequest expects data formatting. I manually set a POST request because the static constructor URL encodes the data, and we're trying to write RAW JSON I believe. I would like to figure out the "proper" way to use the POST constructor, but for now the manual method with the JSON string to a UTF8 byte [] is working for now.