Cleaning up the project
I've removed a lot of the unused scripts and tidied up those that I kept.
I added the ability to toggle the JSON deserializer log output of each field on it's own.
I've removed the update log post for the color index of the materials in the ID color script. I also changed to index math, and hopefully didn't break it.